refactor(model): move discriminator to its own file
