+++ lib/model.js
@@ -7,12 +7,14 @@ var Document = require('./document');
 var DocumentNotFoundError = require('./error/notFound');
 var DivergentArrayError = require('./error').DivergentArrayError;
 var EventEmitter = require('events').EventEmitter;
-var OverwriteModelError = require('./error').OverwriteModelError;
 var PromiseProvider = require('./promise_provider');
 var Query = require('./query');
 var Schema = require('./schema');
 var VersionError = require('./error').VersionError;
 var applyHooks = require('./services/model/applyHooks');
+var applyMethods = require('./services/model/applyMethods');
+var applyStatics = require('./services/model/applyStatics');
+var discriminator = require('./services/model/discriminator');
 var parallel = require('async/parallel');
 var util = require('util');
 var utils = require('./utils');
@@ -820,113 +822,8 @@ Model.prototype.model = function model(name) {
  * @api public
  */
 
-Model.discriminator = function discriminator(name, schema) {
-  var CUSTOMIZABLE_DISCRIMINATOR_OPTIONS = {
-    toJSON: true,
-    toObject: true,
-    _id: true,
-    id: true
-  };
-
-  if (!(schema && schema.instanceOfSchema)) {
-    throw new Error('You must pass a valid discriminator Schema');
-  }
-
-  if (this.schema.discriminatorMapping && !this.schema.discriminatorMapping.isRoot) {
-    throw new Error('Discriminator "' + name +
-        '" can only be a discriminator of the root model');
-  }
-
-  var key = this.schema.options.discriminatorKey;
-  if (schema.path(key)) {
-    throw new Error('Discriminator "' + name +
-        '" cannot have field with name "' + key + '"');
-  }
-
-  function merge(schema, baseSchema) {
-    utils.merge(schema, baseSchema, { retainKeyOrder: true });
-
-    var obj = {};
-    obj[key] = {
-      default: name,
-      set: function(newName) {
-        if (newName === name) {
-          return name;
-        }
-        throw new Error('Can\'t set discriminator key "' + key + '"');
-      }
-    };
-    obj[key][schema.options.typeKey] = String;
-    schema.add(obj);
-    schema.discriminatorMapping = {key: key, value: name, isRoot: false};
-
-    if (baseSchema.options.collection) {
-      schema.options.collection = baseSchema.options.collection;
-    }
-
-    var toJSON = schema.options.toJSON;
-    var toObject = schema.options.toObject;
-    var _id = schema.options._id;
-    var id = schema.options.id;
-
-    var keys = Object.keys(schema.options);
-    schema.options.discriminatorKey = baseSchema.options.discriminatorKey;
-
-    for (var i = 0; i < keys.length; ++i) {
-      var _key = keys[i];
-      if (!CUSTOMIZABLE_DISCRIMINATOR_OPTIONS[_key]) {
-        if (!utils.deepEqual(schema.options[_key], baseSchema.options[_key])) {
-          throw new Error('Can\'t customize discriminator option ' + _key +
-              ' (can only modify ' +
-              Object.keys(CUSTOMIZABLE_DISCRIMINATOR_OPTIONS).join(', ') +
-              ')');
-        }
-      }
-    }
-
-    schema.options = utils.clone(baseSchema.options);
-    if (toJSON) schema.options.toJSON = toJSON;
-    if (toObject) schema.options.toObject = toObject;
-    if (typeof _id !== 'undefined') {
-      schema.options._id = _id;
-    }
-    schema.options.id = id;
-
-    schema.callQueue = baseSchema.callQueue.concat(schema.callQueue.slice(schema._defaultMiddleware.length));
-    schema._requiredpaths = undefined; // reset just in case Schema#requiredPaths() was called on either schema
-  }
-
-  // merges base schema into new discriminator schema and sets new type field.
-  merge(schema, this.schema);
-
-  if (!this.discriminators) {
-    this.discriminators = {};
-  }
-
-  if (!this.schema.discriminatorMapping) {
-    this.schema.discriminatorMapping = {key: key, value: null, isRoot: true};
-  }
-
-  if (this.discriminators[name]) {
-    throw new Error('Discriminator with name "' + name + '" already exists');
-  }
-  if (this.db.models[name]) {
-    throw new OverwriteModelError(name);
-  }
-
-  this.discriminators[name] = this.db.model(name, schema, this.collection.name);
-  this.discriminators[name].prototype.__proto__ = this.prototype;
-  Object.defineProperty(this.discriminators[name], 'baseModelName', {
-    value: this.modelName,
-    configurable: true,
-    writable: false
-  });
-
-  // apply methods and statics
-  applyMethods(this.discriminators[name], schema);
-  applyStatics(this.discriminators[name], schema);
-
-  return this.discriminators[name];
+Model.discriminator = function(name, schema) {
+  return discriminator(this, name, schema);
 };
 
 // Model (class) features
@@ -3403,49 +3300,6 @@ Model.compile = function compile(name, schema, collectionName, connection, base)
   };
 
   return model;
-};
-
-/*!
- * Register methods for this model
- *
- * @param {Model} model
- * @param {Schema} schema
- */
-var applyMethods = function(model, schema) {
-  function apply(method, schema) {
-    Object.defineProperty(model.prototype, method, {
-      get: function() {
-        var h = {};
-        for (var k in schema.methods[method]) {
-          h[k] = schema.methods[method][k].bind(this);
-        }
-        return h;
-      },
-      configurable: true
-    });
-  }
-  for (var method in schema.methods) {
-    if (schema.tree.hasOwnProperty(method)) {
-      throw new Error('You have a method and a property in your schema both ' +
-        'named "' + method + '"');
-    }
-    if (typeof schema.methods[method] === 'function') {
-      model.prototype[method] = schema.methods[method];
-    } else {
-      apply(method, schema);
-    }
-  }
-};
-
-/*!
- * Register statics for this model
- * @param {Model} model
- * @param {Schema} schema
- */
-var applyStatics = function(model, schema) {
-  for (var i in schema.statics) {
-    model[i] = schema.statics[i];
-  }
 };
 
 /*!
