+++ lib/services/model/applyStatics.js
@@ -1,0 +1,12 @@
+'use strict';
+
+/*!
+ * Register statics for this model
+ * @param {Model} model
+ * @param {Schema} schema
+ */
+module.exports = function applyStatics(model, schema) {
+  for (var i in schema.statics) {
+    model[i] = schema.statics[i];
+  }
+};
